base class, file splits
