+++ src/augmentedExp.ts
@@ -1,0 +1,8 @@
+export class AugmentedExp extends RegExp {
+    private _template: string
+
+    constructor(pattern: string, flags: string, template: string) {
+        super(pattern, flags);
+        this._template = template;
+    }
+}
