+++ src/objectCloner.ts
@@ -1,0 +1,5 @@
+export class ObjectCloner {
+    static deepCopy<T extends object>(obj: T) : T {
+        return JSON.parse(JSON.stringify(obj));
+    }
+}
